Utvikling over tid i egen komponent
